+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Upgraded `react-apollo` to major 3.
+
 ## [1.2.1] - 2019-11-07
 
 ### Fixed